+++ src/pages/api/projects.ts
@@ -1,29 +1,6 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-
-const projects = [
-  {
-    name: 'Cool Project Title',
-    description: 'This was a super cool project',
-    url: '/one',
-    highlights: ['thing01', 'thing02', 'thing03'],
-    id: 1,
-  },
-  {
-    name: 'This is another neat one',
-    description: 'Amazingly cool stuff',
-    url: '/two',
-    highlights: ['thing01', 'thing02', 'thing03', 'thing04'],
-    id: 2,
-  },
-  {
-    name: 'Very cool of me!',
-    description: 'I know right',
-    url: '/three',
-    highlights: ['thing01', 'thing02'],
-    id: 3,
-  },
-];
+import getProjects from 'server/service/projects/get';
 
 export default async function handleProjects(_: NextApiRequest, res: NextApiResponse): Promise<void> {
-  res.status(200).json(projects);
+  res.status(200).json(getProjects());
 }
